Move OccMapTree to moveit_core/collision_detection
