+++ moveit_core/collision_detection/include/moveit/collision_detection/occupancy_map.h
@@ -42,7 +42,7 @@
 #include <boost/function.hpp>
 #include <memory>
 
-namespace occupancy_map_monitor
+namespace collision_detection
 {
 typedef octomap::OcTreeNode OccMapNode;
 
@@ -115,4 +115,4 @@ private:
 
 using OccMapTreePtr = std::shared_ptr<OccMapTree>;
 using OccMapTreeConstPtr = std::shared_ptr<const OccMapTree>;
-}  // namespace occupancy_map_monitor
+}  // namespace collision_detection
